% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_sum_of_squares.R
\name{calc_sum_of_squares}
\alias{calc_sum_of_squares}
\title{Calculate sum of squares differences for a vector}
\usage{
calc_sum_of_squares(x)
}
\arguments{
\item{x}{A vector.}
}
\value{
. A number.
}
\description{
Calculate sum of squares differences for a vector
}
\examples{
calc_sum_of_squares(c(1, 2, 3, 4, 5, 6))
}
