% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{packageReview-package}
\alias{packageReview}
\alias{packageReview-package}
\title{packageReview: An R package to demonstrate code reviews using GitHub}
\description{
A minimal R package to provide tutorial on how to conduct code reviews using GitHub's pull request system. The package contains dummy functions and documentation for the purpose of demonstration.
}
\author{
\strong{Maintainer}: Joshua W. Lambert \email{joshua.lambert@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5218-3046}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item James M. Azam \email{james.azam@lshtm.ac.uk} (\href{https://orcid.org/0000-0001-5782-7330}{ORCID}) [contributor, reviewer]
}

}
\keyword{internal}
